refactored ec2metadata.py into cli and lib
